server_app/api/requirements.txt
Fabric==1.4.1
Flask==0.8
Flask-Cache==0.4.0
Flask-Script==0.3.3
Jinja2==2.6
MultipartPostHandler==0.1.0
Werkzeug==0.8.3
argparse==1.2.1
dictshield==0.4.3
http-parser==0.7.5
psycopg2==2.4.5
py-bcrypt==0.2
pycrypto==2.5
python-memcached==1.48
git+git://github.com/dlrust/python-memcached-stats.git
restkit==4.1.2
socketpool==0.3.0
ssh==1.7.13
wsgiref==0.1.2
